ref: badge
refactor `badge` to `BadgeController` which now use `badge` from shadcn
